Use fewer ints

+++ geometry.hpp
@@ -57,10 +57,10 @@ typedef std::vector<draw> drawvec;
 
 drawvec decode_geometry(FILE *meta, long *geompos, int z, unsigned tx, unsigned ty, long *bbox, unsigned initial_x, unsigned initial_y);
 void to_tile_scale(drawvec &geom, int z, int detail);
-drawvec remove_noop(drawvec geom, int type, int shift);
+drawvec remove_noop(drawvec geom, int type, size_t shift);
 drawvec clip_point(drawvec &geom, int z, long buffer);
-drawvec clean_or_clip_poly(drawvec &geom, int z, int buffer, bool clip);
-drawvec simple_clip_poly(drawvec &geom, int z, int buffer);
+drawvec clean_or_clip_poly(drawvec &geom, int z, long buffer, bool clip);
+drawvec simple_clip_poly(drawvec &geom, int z, long buffer);
 drawvec close_poly(drawvec &geom);
 drawvec reduce_tiny_poly(drawvec &geom, int z, int detail, bool *reduced, double *accum_area);
 drawvec clip_lines(drawvec &geom, int z, long buffer);
